@@ -1,14 +1,16 @@
 package com.kamestudio.noticeappmanager;
 
-import android.app.ActivityManager;
-import android.content.Context;
+
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
+import android.util.DisplayMetrics;
 import android.util.Log;
+import android.view.Display;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.FrameLayout;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -16,7 +18,17 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.MobileAds;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.ads.RequestConfiguration;
+import com.google.android.gms.ads.initialization.InitializationStatus;
+import com.google.android.gms.ads.initialization.OnInitializationCompleteListener;
 import com.kamestudio.noticeappmanager.adapter.PackageSetupAdapter;
+import com.kamestudio.noticeappmanager.ads.GoogleMobileAdsConsentManager;
 import com.kamestudio.noticeappmanager.data.DataStoreUtil;
 import com.kamestudio.noticeappmanager.databinding.FragmentFirstBinding;
 import com.kamestudio.noticeappmanager.enity.ItemPackage;
@@ -27,29 +39,143 @@ import com.kamestudio.noticeappmanager.viewmodel.NotificationViewModel;
 import org.greenrobot.eventbus.EventBus;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 public class FirstFragment extends Fragment implements Util{
-
+    private static final String TAG = "FirstFragment";
     private FragmentFirstBinding binding;
     private NotificationViewModel viewModel = null;
     private PackageSetupAdapter adapter;
-    private FirstFragment fragment;
+
+
+    // Ads config
+    private static final String AD_UNIT_ID = "ca-app-pub-3940256099942544/6300978111";
+    private GoogleMobileAdsConsentManager googleMobileAdsConsentManager;
+    private AdView adView;
+    private FrameLayout adContainerView;
+    private AtomicBoolean initialLayoutComplete = new AtomicBoolean(true);
+    private final AtomicBoolean isMobileAdsInitializeCalled = new AtomicBoolean(false);
 
     @Override
     public View onCreateView(
-            LayoutInflater inflater, ViewGroup container,
+            @NonNull LayoutInflater inflater, ViewGroup container,
             Bundle savedInstanceState
     ) {
-
         binding = FragmentFirstBinding.inflate(inflater, container, false);
+        adsConfigSettings();
         return binding.getRoot();
+    }
+
+
+
+    private void adsConfigSettings() {
+        adContainerView = binding.adContainer;
+        googleMobileAdsConsentManager = GoogleMobileAdsConsentManager.getInstance(this.getContext());
+
+        googleMobileAdsConsentManager.gatherConsent(
+                requireActivity(),
+                consentError -> {
+                    if (consentError != null) {
+                        // Consent not obtained in current session.
+                        Log.w(
+                                TAG,
+                                String.format("%s: %s", consentError.getErrorCode(), consentError.getMessage()));
+                    }
+
+                    if (googleMobileAdsConsentManager.canRequestAds()) {
+                        initializeMobileAdsSdk();
+                    }
+
+                    if (googleMobileAdsConsentManager.isPrivacyOptionsRequired()) {
+                        // Regenerate the options menu to include a privacy setting.
+                        requireActivity().invalidateOptionsMenu();
+                    }
+                });
+        // This sample attempts to load ads using consent obtained in the previous session.
+        if (googleMobileAdsConsentManager.canRequestAds()) {
+            initializeMobileAdsSdk();
+        }
+
+        // Since we're loading the banner based on the adContainerView size, we need to wait until this
+        // view is laid out before we can get the width.
+        adContainerView
+                .getViewTreeObserver()
+                .addOnGlobalLayoutListener(
+                        () -> {
+                            if (!initialLayoutComplete.getAndSet(true)
+                                    && googleMobileAdsConsentManager.canRequestAds()) {
+                                loadBanner();
+                            }
+                        });
+
+        // Set your test devices. Check your logcat output for the hashed device ID to
+        // get test ads on a physical device. e.g.
+        // "Use RequestConfiguration.Builder().setTestDeviceIds(Arrays.asList("ABCDEF012345"))
+        // to get test ads on this device."
+        MobileAds.setRequestConfiguration(
+                //test devices
+//                new RequestConfiguration.Builder().setTestDeviceIds(Arrays.asList("446B401A8B1764C39A87843B3BD55F2D")).build());
+                // production devices
+                new RequestConfiguration.Builder().build());
+    }
+
+    private void loadBanner() {
+        // Create a new ad view.
+        adView = new AdView(this.getContext());
+        adView.setAdUnitId(AD_UNIT_ID);
+        adView.setAdSize(getAdSize());
+//        adView.setAdSize(new AdSize(300, 50));
+
+        // Replace ad container with new ad view.
+        adContainerView.removeAllViews();
+        adContainerView.addView(adView);
+
+        // Start loading the ad in the background.
+        AdRequest adRequest = new AdRequest.Builder().build();
+        adView.loadAd(adRequest);
+    }
+
+    private void initializeMobileAdsSdk() {
+        if (isMobileAdsInitializeCalled.getAndSet(true)) {
+            return;
+        }
+
+        // Initialize the Mobile Ads SDK.
+        MobileAds.initialize(
+                this.getContext(),
+                initializationStatus -> {});
+
+        // Load an ad.
+        if (initialLayoutComplete.get()) {
+            loadBanner();
+        }
+    }
+
+    private AdSize getAdSize() {
+        // Determine the screen width (less decorations) to use for the ad width.
+        Display display = requireActivity().getWindowManager().getDefaultDisplay();
+        DisplayMetrics outMetrics = new DisplayMetrics();
+        display.getMetrics(outMetrics);
+
+        float density = outMetrics.density;
+
+        float adWidthPixels = adContainerView.getWidth();
+        Log.d(TAG, "adContainerView -- width: " + adWidthPixels);
+
+        // If the ad hasn't been laid out, default to the full screen width.
+        if (adWidthPixels == 0) {
+            adWidthPixels = outMetrics.widthPixels;
+        }
+
+        int adWidth = (int) (adWidthPixels / density);
+        return AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(this.getContext(), adWidth);
     }
 
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        fragment = this;
-        List<ItemPackage> list = DataStoreUtil.getInstance(requireActivity()).getPackages();
+        List<ItemPackage> list = DataStoreUtil.getInstance(getActivity()).getPackages();
         viewModel = new ViewModelProvider(requireActivity()).get(NotificationViewModel.class);
         //observable data changing from other fragment
         viewModel.getListMutableLiveData().observe(getViewLifecycleOwner(), itemPackages -> {
@@ -58,42 +184,36 @@ public class FirstFragment extends Fragment implements Util{
                 updateUI(itemPackages);
         });
 
-        viewModel.setListMutableLiveData(list);
-        Log.d(TAG, "onViewCreated: DataStore" + list);
+        viewModel.getListMutableLiveData().setValue(list);
+        Log.d(TAG, "onViewCreated: DataStore " + list);
         Log.d(TAG, "onViewCreated: " + viewModel.getListMutableLiveData().getValue());
         bindingViewFunction();
-        startNoticeService();
+        if (!NoticeService.IS_RUNNING){
+            startNoticeService();
+        }
     }
 
     private void bindingViewFunction(){
-        //start forceground service
-        binding.buttonStartService.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startNoticeService();
+        //start foreground service
+        binding.buttonStartService.setOnClickListener(view -> startNoticeService());
+
+        binding.buttonStopService.setOnClickListener(view -> {
+            try {
+                Toast.makeText(getContext(), "Stop Service", Toast.LENGTH_SHORT).show();
+                Intent intent = new Intent(getActivity(), NoticeService.class);
+                intent.setAction(FOREGROUND_STOP_ACTION);
+                getActivity().startService(intent);
+                EventBus.getDefault().post(new MessageEvent(NoticeService.ACTION_STOP_SERVICE));
+                Log.d(TAG, "onClick: buttonStopService ");
+            }
+            catch (Exception ex){
+                Log.e(TAG, "onClick: buttonStopService" + ex);
             }
         });
 
-        binding.buttonStopService.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                try {
-                    Toast.makeText(getContext(), "Stop Service", Toast.LENGTH_SHORT).show();
-                    EventBus.getDefault().post(new MessageEvent(NoticeService.ACTION_STOP_SERVICE));
-                    Log.d(TAG, "onClick: buttonStopService ");
-                }
-                catch (Exception ex){
-                    Log.e(TAG, "onClick: buttonStopService" + ex.toString());
-                }
-            }
-        });
-
-        binding.buttonSettingNotify.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Toast.makeText(getContext(), "Stop Notify", Toast.LENGTH_SHORT).show();
-                startActivity(new Intent(((MainActivity) getActivity()).ACTION_NOTIFICATION_LISTENER_SETTINGS));
-            }
+        binding.buttonSettingNotify.setOnClickListener(view -> {
+//            Toast.makeText(getContext(), "Stop Notify", Toast.LENGTH_SHORT).show();
+            startActivity(new Intent(((MainActivity) getActivity()).ACTION_NOTIFICATION_LISTENER_SETTINGS));
         });
 
         //create view of item package
@@ -105,10 +225,10 @@ public class FirstFragment extends Fragment implements Util{
     }
 
     public void updateUI(List<ItemPackage> itemPackages){
-        String data = "";
+        StringBuilder data = new StringBuilder();
         if(itemPackages != null && itemPackages.size() != 0) {
             for (ItemPackage item : itemPackages) {
-                data += item.getPackageInfo().packageName + "\n";
+                data.append(item.getPackageInfoName()).append("\n");
             }
         }
         else{
@@ -118,36 +238,50 @@ public class FirstFragment extends Fragment implements Util{
         adapter.setItemPackageList(itemPackages);
         adapter.notifyDataSetChanged();
         DataStoreUtil.getInstance(this.requireContext()).setPackages(viewModel.getListMutableLiveData().getValue());
-        Log.d("Notify App", "updateUI: "+data);
+        Log.d(TAG, "updateUI: "+data);
     }
 
     @Override
     public void onDestroyView() {
-//        DataStoreUtil.getInstance(this.requireContext()).setPackages(viewModel.getListMutableLiveData().getValue());
         super.onDestroyView();
         binding = null;
     }
 
-    private void startNoticeService(){
-        Intent intent = new Intent(getActivity(), NoticeService.class);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            getActivity().startForegroundService(intent);
-            Toast.makeText(getContext(), "Start Service 1", Toast.LENGTH_SHORT).show();
-        }
-        else{
-            getActivity().startService(intent);
-            Toast.makeText(getContext(), "Start Service 2", Toast.LENGTH_SHORT).show();
+    @Override
+    public void onResume() {
+        super.onResume();
+        if (adView != null) {
+            adView.resume();
         }
     }
 
-    private boolean isServiceRunning(Context context, Class<?> serviceClass) {
-        ActivityManager manager = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
-        for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
-                return true;
-            }
+    @Override
+    public void onPause() {
+        if (adView != null) {
+            adView.resume();
         }
-        return false;
+        super.onPause();
+    }
+
+    @Override
+    public void onDestroy() {
+        if (adView != null) {
+            adView.destroy();
+        }
+        super.onDestroy();
+    }
+
+    private void startNoticeService(){
+        Intent intent = new Intent(getActivity(), NoticeService.class);
+        intent.setAction(FOREGROUND_START_ACTION);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            getActivity().startForegroundService(intent);
+//            Toast.makeText(getContext(), "Start Service 1", Toast.LENGTH_SHORT).show();
+        }
+        else{
+            getActivity().startService(intent);
+//            Toast.makeText(getContext(), "Start Service 2", Toast.LENGTH_SHORT).show();
+        }
     }
 
 }
