@@ -10,18 +10,18 @@ import com.google.android.gms.ads.initialization.OnInitializationCompleteListene
 
 public class NoticeApplication extends Application {
 
-    private static AppOpenManager appOpenManager;
+//    private static AppOpenManager appOpenManager;
     private static NoticeApplication instance;
 
     @Override
     public void onCreate() {
         super.onCreate();
-        MobileAds.initialize(this, new OnInitializationCompleteListener() {
-            @Override
-            public void onInitializationComplete(@NonNull InitializationStatus initializationStatus) {
-            }
-        });
-        appOpenManager = new AppOpenManager(this);
+//        MobileAds.initialize(this, new OnInitializationCompleteListener() {
+//            @Override
+//            public void onInitializationComplete(@NonNull InitializationStatus initializationStatus) {
+//            }
+//        });
+//        appOpenManager = new AppOpenManager(this);
     }
 
     public NoticeApplication getInstance() {
